moved hmc, mala, smc, and metropolis into bayes_kit.algorithms module
