+++ bayes_kit/__init__.py
@@ -6,11 +6,13 @@ from .rhat import rhat
 
 # classes
 from .ensemble import Stretcher
-from .hmc import HMCDiag
-from .mala import MALA
-from .metropolis import Metropolis, MetropolisHastings
+from .algorithms import \
+    HMCDiag, \
+    MALA, \
+    Metropolis, \
+    MetropolisHastings, \
+    TemperedLikelihoodSMC
 from .ensemble import Stretcher
-from .smc import TemperedLikelihoodSMC
 
 __all__ = [
     "HMCDiag",
