+++ bayes_kit/algorithms/hmc.py
@@ -2,7 +2,7 @@ from typing import Iterator, Optional, Union
 from numpy.typing import NDArray
 import numpy as np
 
-from .model_types import GradModel
+from bayes_kit.model_types import GradModel
 
 Draw = tuple[NDArray[np.float64], float]
